// lib/api/generated/model/trust.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Trust Ledger API
 * Track money that other people trust you with temporarily.

## Authentication
All endpoints require a valid OAuth2 token.
Pass it as: `Authorization: Bearer <token>`
 * OpenAPI spec version: 1.0.0
 */
import type { Currency } from './currency';
import type { Person } from './person';
import type { StatusEnum } from './statusEnum';

export interface Trust {
  readonly id: string;
  readonly person: Person;
  /** @pattern ^-?\d{0,10}(?:\.\d{0,2})?$ */
  amount: string;
  readonly currency: Currency;
  status?: StatusEnum;
  notes?: string;
  readonly created_by: string;
  readonly updated_by: string;
  readonly created_at: string;
  readonly updated_at: string;
}